% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getMus}
\alias{getMus}
\title{Get point-estimates of prognostic effect (in sample)}
\usage{
getMus(fit)
}
\arguments{
\item{fit}{An XBCF fit object.}
}
\value{
An array of prognostic effect point estimates.
}
\description{
Get point-estimates of prognostic effect (in sample)
}
