% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getTaus}
\alias{getTaus}
\title{Get point-estimates of treatment effect (in sample)}
\usage{
getTaus(fit)
}
\arguments{
\item{fit}{An XBCF fit object.}
}
\value{
An array of treatment effect point estimates.
}
\description{
Get point-estimates of treatment effect (in sample)
}
