% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.XBART.R
\name{predict.XBART}
\alias{predict.XBART}
\title{Predicting new observations using fitted XBART regression model.}
\usage{
\method{predict}{XBART}(object, X, ...)
}
\arguments{
\item{object}{Fitted \eqn{object} returned from XBART function.}

\item{X}{A matrix of input testing data \eqn{X}}
}
\value{
yhats A vector of predictted outcome \eqn{Y} for the testing data.
}
\description{
This function predicts testing data given fitted XBART regression model.
}
\details{
XBART draws multiple samples of the forests (sweeps), each forest is an ensemble of trees. The final prediction is taking sum of trees in each forest, and average across different sweeps (without burnin sweeps).
}
