% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.XBART.R
\name{predict.XBCFdiscrete}
\alias{predict.XBCFdiscrete}
\title{Predicting new observations using fitted XBCF binary treatment model.}
\usage{
\method{predict}{XBCFdiscrete}(object, X_con, X_mod, Z, pihat = NULL, burnin = 0L, ...)
}
\arguments{
\item{object}{Fitted \eqn{object} returned from XBART function.}

\item{X_con}{A matrix of input testing data for the prognostic forest.}

\item{X_mod}{A matrix of input testing data for the treatment forest.}

\item{Z}{A vector of input testing data for the treatment variable.}

\item{pihat}{An array of propensity score estimates.}

\item{burnin}{The number of burn-in iterations to discard from averaging (the default value is 0).}
}
\value{
A list containing predicted prognostic term, treatment effect and final outcome \eqn{Y}.
}
\description{
This function predicts testing data given fitted XBCF continuous treatment model.
}
\details{
XBCF draws multiple samples of the forests (sweeps), each forest is an ensemble of trees. The final prediction returns predicted prognostic term, treatment effect and the final outcome \eqn{Y}
}
\seealso{
\link[XBART]{XBCF.discrete}
}
