\name{XBART-package}
\alias{XBART-package}
\alias{XBART}
\docType{package}
\title{
  XBART: Accelerated Bayesian Additive Regression Trees
}
\description{
  A stochastic hill climbing algorithm for Bayesian additive regression trees model. 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Jingyu He, Saar Yalov and P. Richard Hahn

Maintainer: Jingyu He <jingyu.he@chicagobooth.edu>
}
\references{
  XBART: Accelerated Bayesian Additive Regression Trees, AISTATS 2019.
}
\keyword{ package }
\seealso{
  XBART
}

