% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBCF.continuous.R
\name{XBCF.continuous}
\alias{XBCF.continuous}
\title{XBCF causal forest for continuous treatment variable.}
\usage{
XBCF.continuous(
  y,
  Z,
  X_con,
  X_mod,
  num_trees_con,
  num_trees_mod,
  num_sweeps,
  max_depth = 250,
  Nmin = 1,
  num_cutpoints = 100,
  alpha_con = 0.95,
  beta_con = 1.25,
  alpha_mod = 0.95,
  beta_mod = 1.25,
  tau_con = NULL,
  tau_mod = NULL,
  no_split_penality = NULL,
  burnin = 1L,
  mtry_con = NULL,
  mtry_mod = NULL,
  p_categorical_con = 0L,
  p_categorical_mod = 0L,
  kap = 16,
  s = 4,
  tau_con_kap = 3,
  tau_con_s = 0.5,
  tau_mod_kap = 3,
  tau_mod_s = 0.5,
  verbose = FALSE,
  update_tau = TRUE,
  parallel = TRUE,
  random_seed = NULL,
  sample_weights = TRUE,
  nthread = 0,
  ...
)
}
\arguments{
\item{y}{A vector of outcome variable of length n, expected to be continuous.}

\item{Z}{A vector of treatment variable of length n, expected to be continuous.}

\item{X_con}{A matrix of input for the prognostic forest of size n by p_con. Column order matters: continuous features should all go before categorical. The number of categorical variables is p_categorical_con.}

\item{X_mod}{A matrix of input for the treatment forest of size n by p_con. Column order matters: continuous features should all go before categorical. The number of categorical variables is p_categorical_mod.}

\item{num_trees_con}{Integer, number of trees in the prognostic forest.}

\item{num_trees_mod}{Integer, number of trees in the treatment forest.}

\item{num_sweeps}{Integer, number of sweeps to fit for both forests.}

\item{max_depth}{Integer, maximum depth of the trees. The tree will stop grow if reaches the max depth.}

\item{Nmin}{Integer, minimal number of data points in a leaf. Any leaf will stop split if number of data points within is smaller than Nmin.}

\item{num_cutpoints}{Integer, number of cutpoint candidates to consider for each variable. Take in quantiles of the data.}

\item{alpha_con}{Scalar, BART prior parameter for prognostic forest. The default value is 0.95.}

\item{beta_con}{Scalar, BART prior parameter for prognostic forest. The default value is 1.25.}

\item{alpha_mod}{Scalar, BART prior parameter for treatment forest. The default value is 0.95.}

\item{beta_mod}{Scalar, BART prior parameter for treatment forest. The default value is 1.25.}

\item{tau_con}{Scalar, prior parameter for prognostic forest. The default value is 0.6 * var(y) / num_trees_con.}

\item{tau_mod}{Scalar, prior parameter for treatment forest. The default value is 0.1 * var(y) / num_trees_mod.}

\item{no_split_penality}{Weight of no-split option. The default value is log(num_cutpoints), or you can take any other number in log scale.}

\item{burnin}{Integer, number of burnin sweeps.}

\item{mtry_con}{Integer, number of X variables to sample at each split of the prognostic forest.}

\item{mtry_mod}{Integer, number of X variables to sample at each split of the treatment forest.}

\item{p_categorical_con}{Integer, number of categorical variables in X_con, note that all categorical variables should be put after continuous variables. Default value is 0.}

\item{p_categorical_mod}{Integer, number of categorical variables in X_mod, note that all categorical variables should be put after continuous variables. Default value is 0.}

\item{kap}{Scalar, parameter of the inverse gamma prior on residual variance sigma^2. Default value is 16.}

\item{s}{Scalar, parameter of the inverse gamma prior on residual variance sigma^2. Default value is 4.}

\item{tau_con_kap}{Scalar, parameter of the inverse gamma prior on tau_con. Default value is 3.}

\item{tau_con_s}{Scalar, parameter of the inverse gamma prior on tau_con. Default value is 0.5.}

\item{tau_mod_kap}{Scalar, parameter of the inverse gamma prior on tau_mod. Default value is 3.}

\item{tau_mod_s}{Scalar, parameter of the inverse gamma prior on tau_mod. Default value is 0.5.}

\item{verbose}{Bool, whether to print fitting process on the screen or not.}

\item{update_tau}{Bool. If TRUE, update the prior of leaf mean.}

\item{random_seed}{Integer, random seed for replication.}

\item{sample_weights}{Bool, if TRUE, the weight to sample \eqn{X} variables at each tree will be sampled.}

\item{nthread}{Integer, number of threads to use if run in parallel.}

\item{paralll}{Bool, whether to run in parallel on multiple CPU threads.}
}
\value{
A list contains fitted trees as well as parameter draws at each sweep.
}
\description{
This function fits XBCF causal forest model with continuous treatment variable. In sepecific, the model is \eqn{y = \tau(X_{con}) + \mu(X_{mod}) \times z + \epsilon, \quad \epsilon\sim N(0, \sigma^2)}. Where the \eqn{\tau(X_{con})} and \eqn{\mu(X_{mod})} are two XBART forest called prognostic and treatment forest respectively.
}
