\name{XBART-package}
\alias{XBART-package}
\docType{package}
\title{
  Accelerated Bayesian Additive Regression Trees (XBART)
}
\description{
  A stochastic hill climbing algorithm for Bayesian additive regression trees model. 
}
\details{
  This pacakge contains the XBART algorithm for both regression and classification tasks. Furthermore, it fits XBCF causal forest model with continuous treatment effects. It also implements the Gaussian extrapolation algorithm for prediction if the testing data is out of the range of the training.
}
\author{
Jingyu He, Saar Yalov, P. Richard Hahn and Meijiang Wang

Maintainer: Jingyu He <jingyuhe@cityu.edu.hk>
}
\references{
  XBART: Accelerated Bayesian Additive Regression Trees, AISTATS 2019.
}
\keyword{ package }
\seealso{
  XBART, XBART.mutlinomial, XBCF.continuous
}

