\name{XBART-package}
\alias{XBART-package}
\docType{package}
\title{
  Accelerated Bayesian Additive Regression Trees (XBART)
}
\description{
  A stochastic hill climbing algorithm for Bayesian additive regression trees model. 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Jingyu He, Saar Yalov, P. Richard Hahn and Meijiang Wang

Maintainer: Jingyu He <jingyuhe@cityu.edu.hk>
}
\references{
  XBART: Accelerated Bayesian Additive Regression Trees, AISTATS 2019.
}
\keyword{ package }
\seealso{
  XBART
}

