% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBART.R
\name{XBART}
\alias{XBART}
\title{XBART main function of XBART regression.}
\usage{
XBART(
  y,
  X,
  num_trees,
  num_sweeps,
  max_depth = 250,
  Nmin = 1,
  num_cutpoints = 100,
  alpha = 0.95,
  beta = 1.25,
  tau = NULL,
  no_split_penality = NULL,
  burnin = 1L,
  mtry = NULL,
  p_categorical = 0L,
  kap = 16,
  s = 4,
  tau_kap = 3,
  tau_s = 0.5,
  verbose = FALSE,
  update_tau = TRUE,
  parallel = TRUE,
  random_seed = NULL,
  sample_weights = TRUE,
  nthread = 0,
  ...
)
}
\arguments{
\item{y}{A vector of outcome variable of length n, expected to be continuous.}

\item{X}{A matrix of input for the tree of size n by p. Column order matters: continuous features should all go before categorical. The number of categorical variables is p_categorical_con.}

\item{num_trees}{Integer, number of trees in the prognostic forest.}

\item{num_sweeps}{Integer, number of sweeps to fit for both forests.}

\item{max_depth}{Integer, maximum depth of the trees. The tree will stop grow if reaches the max depth.}

\item{Nmin}{Integer, minimal number of data points in a leaf. Any leaf will stop split if number of data points within is smaller than Nmin.}

\item{num_cutpoints}{Integer, number of cutpoint candidates to consider for each variable. Take in quantiles of the data.}

\item{alpha}{Scalar, BART prior parameter for trees. The default value is 0.95.}

\item{beta}{Scalar, BART prior parameter for trees. The default value is 1.25.}

\item{tau}{Scalar, prior parameter for the trees. The default value is 1 / num_trees.}

\item{no_split_penality}{Weight of no-split option. The default value is log(num_cutpoints), or you can take any other number in log scale.}

\item{burnin}{Integer, number of burnin sweeps.}

\item{mtry}{Integer, number of X variables to sample at each split of the tree.}

\item{p_categorical}{Integer, number of categorical variables in X, note that all categorical variables should be put after continuous variables. Default value is 0.}

\item{kap}{Scalar, parameter of the inverse gamma prior on residual variance sigma^2. Default value is 16.}

\item{s}{Scalar, parameter of the inverse gamma prior on residual variance sigma^2. Default value is 4.}

\item{tau_kap}{Scalar, parameter of the inverse gamma prior on tau. Default value is 3.}

\item{tau_s}{Scalar, parameter of the inverse gamma prior on tau. Default value is 0.5.}

\item{verbose}{Bool, whether to print fitting process on the screen or not.}

\item{update_tau}{Bool, if TRUE, update the prior of leaf mean.}

\item{random_seed}{Integer, random seed for replication.}

\item{sample_weights}{Bool, if TRUE, the weight to sample \eqn{X} variables at each tree will be sampled.}

\item{nthread}{Integer, number of threads to use if run in parallel.}

\item{paralll}{Bool, whether to run in parallel on multiple CPU threads.}
}
\value{
A list contains fitted trees as well as parameter draws at each sweep.
}
\description{
XBART main function of XBART regression.
}
